[FIX] Type conversion error.

+++ server/channel/src/CharacterManager.cpp
@@ -5058,7 +5058,7 @@ bool CharacterManager::UpdateExperience(const std::shared_ptr<
 
             int32_t points = 0;
             int32_t bonusCount = 0;
-            for(int8_t lvl = startingLevel + 1; lvl <= level; lvl++)
+            for(int8_t lvl = (int8_t)(startingLevel + 1); lvl <= level; lvl++)
             {
                 points = points + (int32_t)((lvl / 5) + 2);
